cleanup and reorganize codegen compiler
