+++ codegen-compiler/src/main/kotlin/com/freeletics/khonshu/codegen/codegen/ActivityGenerator.kt
@@ -1,20 +1,14 @@
-package com.freeletics.khonshu.codegen.codegen.compose
+package com.freeletics.khonshu.codegen.codegen
 
 import com.freeletics.khonshu.codegen.BaseData
 import com.freeletics.khonshu.codegen.NavHostActivityData
-import com.freeletics.khonshu.codegen.codegen.Generator
-import com.freeletics.khonshu.codegen.codegen.common.composableName
-import com.freeletics.khonshu.codegen.codegen.common.retainedComponentClassName
-import com.freeletics.khonshu.codegen.codegen.common.retainedComponentFactoryCreateName
-import com.freeletics.khonshu.codegen.codegen.common.retainedParentComponentClassName
-import com.freeletics.khonshu.codegen.codegen.common.retainedParentComponentGetterName
-import com.freeletics.khonshu.codegen.codegen.util.bundle
-import com.freeletics.khonshu.codegen.codegen.util.getComponent
-import com.freeletics.khonshu.codegen.codegen.util.lateinitPropertySpec
-import com.freeletics.khonshu.codegen.codegen.util.navHost
-import com.freeletics.khonshu.codegen.codegen.util.optInAnnotation
-import com.freeletics.khonshu.codegen.codegen.util.propertyName
-import com.freeletics.khonshu.codegen.codegen.util.setContent
+import com.freeletics.khonshu.codegen.util.bundle
+import com.freeletics.khonshu.codegen.util.getComponent
+import com.freeletics.khonshu.codegen.util.lateinitPropertySpec
+import com.freeletics.khonshu.codegen.util.navHost
+import com.freeletics.khonshu.codegen.util.optInAnnotation
+import com.freeletics.khonshu.codegen.util.propertyName
+import com.freeletics.khonshu.codegen.util.setContent
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier.OVERRIDE
 import com.squareup.kotlinpoet.TypeSpec
