+++ codegen-compiler/src/main/kotlin/com/freeletics/khonshu/codegen/codegen/ActivityModuleGenerator.kt
@@ -1,17 +1,16 @@
-package com.freeletics.khonshu.codegen.codegen.compose
+package com.freeletics.khonshu.codegen.codegen
 
 import com.freeletics.khonshu.codegen.BaseData
-import com.freeletics.khonshu.codegen.codegen.Generator
-import com.freeletics.khonshu.codegen.codegen.util.composeDestination
-import com.freeletics.khonshu.codegen.codegen.util.contributesToAnnotation
-import com.freeletics.khonshu.codegen.codegen.util.deepLinkHandler
-import com.freeletics.khonshu.codegen.codegen.util.deepLinkPrefix
-import com.freeletics.khonshu.codegen.codegen.util.immutableSet
-import com.freeletics.khonshu.codegen.codegen.util.jvmSuppressWildcards
-import com.freeletics.khonshu.codegen.codegen.util.module
-import com.freeletics.khonshu.codegen.codegen.util.multibinds
-import com.freeletics.khonshu.codegen.codegen.util.provides
-import com.freeletics.khonshu.codegen.codegen.util.toImmutableSet
+import com.freeletics.khonshu.codegen.util.composeDestination
+import com.freeletics.khonshu.codegen.util.contributesToAnnotation
+import com.freeletics.khonshu.codegen.util.deepLinkHandler
+import com.freeletics.khonshu.codegen.util.deepLinkPrefix
+import com.freeletics.khonshu.codegen.util.immutableSet
+import com.freeletics.khonshu.codegen.util.jvmSuppressWildcards
+import com.freeletics.khonshu.codegen.util.module
+import com.freeletics.khonshu.codegen.util.multibinds
+import com.freeletics.khonshu.codegen.util.provides
+import com.freeletics.khonshu.codegen.util.toImmutableSet
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier.ABSTRACT
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
