+++ codegen-compiler/src/main/kotlin/com/freeletics/khonshu/codegen/codegen/FileGenerator.kt
@@ -1,70 +1,50 @@
 package com.freeletics.khonshu.codegen.codegen
 
 import com.freeletics.khonshu.codegen.BaseData
-import com.freeletics.khonshu.codegen.ComposeScreenData
+import com.freeletics.khonshu.codegen.NavDestinationData
 import com.freeletics.khonshu.codegen.NavHostActivityData
-import com.freeletics.khonshu.codegen.codegen.common.ComponentGenerator
-import com.freeletics.khonshu.codegen.codegen.common.ComponentProviderGenerator
-import com.freeletics.khonshu.codegen.codegen.common.InnerComposableGenerator
-import com.freeletics.khonshu.codegen.codegen.common.ModuleGenerator
-import com.freeletics.khonshu.codegen.codegen.common.NavDestinationModuleGenerator
-import com.freeletics.khonshu.codegen.codegen.compose.ActivityGenerator
-import com.freeletics.khonshu.codegen.codegen.compose.ActivityModuleGenerator
-import com.freeletics.khonshu.codegen.codegen.compose.OuterComposableGenerator
 import com.squareup.kotlinpoet.FileSpec
 
 public class FileGenerator {
 
     public fun generate(data: BaseData): FileSpec {
         return when (data) {
-            is ComposeScreenData -> generate(data)
+            is NavDestinationData -> generate(data)
             is NavHostActivityData -> generate(data)
         }
     }
 
-    public fun generate(data: ComposeScreenData): FileSpec {
+    public fun generate(data: NavDestinationData): FileSpec {
         val component = ComponentGenerator(data)
-        val module = ModuleGenerator(data)
-        val outerComposable = OuterComposableGenerator(data)
-        val innerComposable = InnerComposableGenerator(data)
+        val componentProvider = NavDestinationComponentProviderGenerator(data)
+        val module = ComponentModuleGenerator(data)
+        val destinationComposable = NavDestinationComposableGenerator(data)
+        val componentComposable = ComponentComposableGenerator(data)
+        val destinationModule = NavDestinationModuleGenerator(data)
 
         return FileSpec.builder(data.packageName, "Khonshu${data.baseName}")
             .addType(component.generate())
-            .addComponentProviderType(data)
+            .addType(componentProvider.generate())
             .addType(module.generate())
-            .addFunction(outerComposable.generate())
-            .addFunction(innerComposable.generate())
-            .addNavDestinationTypes(data)
+            .addFunction(destinationComposable.generate())
+            .addFunction(componentComposable.generate())
+            .addType(destinationModule.generate())
             .build()
     }
 
     public fun generate(data: NavHostActivityData): FileSpec {
         val component = ComponentGenerator(data)
-        val module = ModuleGenerator(data)
+        val module = ComponentModuleGenerator(data)
         val activityModule = ActivityModuleGenerator(data)
         val activity = ActivityGenerator(data)
-        val innerComposable = InnerComposableGenerator(data)
+        val componentComposable = ComponentComposableGenerator(data)
 
         return FileSpec.builder(data.packageName, "Khonshu${data.baseName}")
             .addType(component.generate())
             .addType(module.generate())
             .addType(activityModule.generate())
             .addType(activity.generate())
-            .addFunction(innerComposable.generate())
+            .addFunction(componentComposable.generate())
             .build()
-    }
-
-    private fun FileSpec.Builder.addComponentProviderType(data: BaseData) = apply {
-        if (data.navigation != null) {
-            val componentProvider = ComponentProviderGenerator(data)
-            addType(componentProvider.generate())
-        }
-    }
-
-    private fun FileSpec.Builder.addNavDestinationTypes(data: BaseData) = apply {
-        if (data.navigation != null) {
-            val navDestinationModule = NavDestinationModuleGenerator(data)
-            addType(navDestinationModule.generate())
-        }
     }
 }
